move account sidebar into sep component
